# worldTuner-api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "worldTuner API",
    "version": "1.0.0",
    "description": "统一新库 API，支持 Cloudflare D1 和 SQLite。仅 /api 路径，无旧表或旧接口兼容。匿名游客，无用户体系；拒绝未知/重复参数。所有实体 ID 是 19 位字符串。"
  },
  "servers": [
    {
      "url": "http://localhost:8787/api"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "health",
        "description": "存活探测，不检查数据库。",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string"
                        },
                        "service": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "无效参数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "请求超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误或数据源不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/stations": {
      "get": {
        "operationId": "listStations",
        "description": "公开电台，按票数倒序和 ID 升序分页；offset 不超过 100000。分类按关联 ID 精确匹配，keyword 字面搜索名称、地点、国家、标签和语种。",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10000,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "tagsId",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EntityId"
            }
          },
          {
            "name": "languagesId",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EntityId"
            }
          },
          {
            "name": "countriesId",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EntityId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/StationPage"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "无效参数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "请求超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误或数据源不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/stations/{id}": {
      "get": {
        "operationId": "stationDetail",
        "description": "隐藏、不可用和不存在的电台均为 404。HTTP(S) 流优先非失败，其次首选，最后 ID；默认 URL 与 streams 第一项一致。",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Station"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "streams": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Stream"
                              }
                            },
                            "links": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "$ref": "#/components/schemas/EntityId"
                                  },
                                  "platform": {
                                    "type": "string"
                                  },
                                  "url": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "无效参数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "请求超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "电台不可见或不存在"
          },
          "500": {
            "description": "内部错误或数据源不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/map/stations": {
      "get": {
        "operationId": "mapStations",
        "description": "ID 升序游标分页，排除隐藏、不可用及无坐标记录；点位不聚合，nextCursor 为 null 表示完成。",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5000,
              "default": 2000
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EntityId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "id",
                              "name",
                              "geoLat",
                              "geoLong"
                            ],
                            "properties": {
                              "id": {
                                "$ref": "#/components/schemas/EntityId"
                              },
                              "name": {
                                "type": "string"
                              },
                              "geoLat": {
                                "type": "number",
                                "minimum": -90,
                                "maximum": 90
                              },
                              "geoLong": {
                                "type": "number",
                                "minimum": -180,
                                "maximum": 180
                              }
                            }
                          }
                        },
                        "nextCursor": {
                          "type": "string",
                          "pattern": "^[0-9]{19}$",
                          "example": "0363513228099059713",
                          "description": "19 位十进制字符串，保留前导零，禁止转换为 Number。",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "无效参数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "请求超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误或数据源不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tags": {
      "get": {
        "operationId": "listTags",
        "description": "按公开电台关联数量倒序、ID 升序排列；只返回有公开电台的分类。标签 q 按字面子串匹配。",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 30
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100000,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CatalogItem"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "无效参数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "请求超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误或数据源不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/countries": {
      "get": {
        "operationId": "listCountries",
        "description": "按公开电台关联数量倒序、ID 升序排列；只返回有公开电台的分类。标签 q 按字面子串匹配。",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CatalogItem"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "无效参数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "请求超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误或数据源不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/languages": {
      "get": {
        "operationId": "listLanguages",
        "description": "按公开电台关联数量倒序、ID 升序排列；只返回有公开电台的分类。标签 q 按字面子串匹配。",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CatalogItem"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "无效参数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "请求超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误或数据源不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/metrics/visit": {
      "post": {
        "operationId": "recordAnonymousVisit",
        "tags": [
          "metrics"
        ],
        "description": "Records one page view. The server derives UTC day/month hashes and stores no raw visitor ID. No account or authentication is involved; the result is approximate usage analytics.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "visitorId",
                  "source",
                  "page"
                ],
                "properties": {
                  "visitorId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Random anonymous UUID stored only on the client."
                  },
                  "source": {
                    "type": "string",
                    "enum": [
                      "web",
                      "android"
                    ]
                  },
                  "page": {
                    "type": "string",
                    "enum": [
                      "welcome",
                      "map",
                      "discover",
                      "search",
                      "charts",
                      "settings",
                      "vip",
                      "player"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Visit accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "recorded": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unexpected field"
          },
          "413": {
            "description": "Request body exceeds 1024 characters"
          },
          "415": {
            "description": "JSON content type required"
          },
          "429": {
            "description": "Metrics write rate limit exceeded"
          }
        }
      }
    },
    "/map/snapshot": {
      "get": {
        "operationId": "mapSnapshot",
        "description": "全部公开且有有效坐标的电台，不分页、不聚合。四字段元组，坐标保留五位小数。默认 TTL 300 秒，支持 gzip、弱 ETag；不接受查询参数。",
        "parameters": [
          {
            "in": "header",
            "name": "If-None-Match",
            "schema": {
              "type": "string"
            },
            "description": "上次 ETag；支持多标签与 *。"
          },
          {
            "in": "header",
            "name": "Accept-Encoding",
            "schema": {
              "type": "string"
            },
            "description": "接受 gzip 时压缩，gzip;q=0 返回 identity。"
          }
        ],
        "responses": {
          "200": {
            "description": "全量点位；ETag 校验和剩余缓存 TTL。",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              },
              "Vary": {
                "schema": {
                  "type": "string"
                }
              },
              "Content-Encoding": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip"
                  ]
                },
                "description": "仅 gzip 正文包含此头。"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "count",
                        "points"
                      ],
                      "properties": {
                        "count": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "points": {
                          "type": "array",
                          "items": {
                            "type": "array",
                            "minItems": 4,
                            "maxItems": 4,
                            "description": "固定顺序：[19 位字符串 ID, name, longitude, latitude]；OAS 3.0 不支持位置类型校验。",
                            "items": {
                              "oneOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            },
                            "example": [
                              "0363513228099059713",
                              "Radio",
                              -87.90647,
                              43.0389
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "304": {
            "description": "内容未改变，无正文；复用现有快照并更新缓存有效期。",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "不接受查询参数。"
          },
          "429": {
            "description": "读取限流。"
          },
          "503": {
            "description": "数据源暂时不可用，不缓存失败响应。"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EntityId": {
        "type": "string",
        "pattern": "^[0-9]{19}$",
        "example": "0363513228099059713",
        "description": "19 位十进制字符串，保留前导零，禁止转换为 Number。"
      },
      "Station": {
        "type": "object",
        "required": [
          "id",
          "name",
          "url",
          "urlResolved",
          "geoLat",
          "geoLong",
          "catalogStatus"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/EntityId"
          },
          "name": {
            "type": "string"
          },
          "countryId": {
            "type": "string",
            "pattern": "^[0-9]{19}$",
            "example": "0363513228099059713",
            "description": "19 位十进制字符串，保留前导零，禁止转换为 Number。",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "favicon": {
            "type": "string",
            "nullable": true
          },
          "place": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "countrycode": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "urlResolved": {
            "type": "string",
            "nullable": true
          },
          "codec": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "languagecodes": {
            "type": "string",
            "nullable": true
          },
          "geoLat": {
            "type": "number",
            "nullable": true
          },
          "geoLong": {
            "type": "number",
            "nullable": true
          },
          "votes": {
            "type": "number",
            "nullable": true
          },
          "clickcount": {
            "type": "number",
            "nullable": true
          },
          "bitrate": {
            "type": "number",
            "nullable": true
          },
          "sourceType": {
            "type": "string",
            "enum": [
              "radio_browser",
              "radio_garden",
              "both"
            ]
          },
          "catalogStatus": {
            "type": "string",
            "enum": [
              "active",
              "unverified"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "hls": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "nullable": true
          },
          "lastcheckok": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "nullable": true
          }
        }
      },
      "StationPage": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Station"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "hasMore": {
            "type": "boolean"
          },
          "nextPage": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "Stream": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/EntityId"
          },
          "url": {
            "type": "string"
          },
          "urlResolved": {
            "type": "string",
            "nullable": true
          },
          "codec": {
            "type": "string",
            "nullable": true
          },
          "bitrate": {
            "type": "number",
            "nullable": true
          },
          "hls": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "nullable": true
          },
          "isPrimary": {
            "type": "boolean"
          },
          "lastcheckok": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "nullable": true
          },
          "lastCheckedAt": {
            "type": "string",
            "nullable": true
          },
          "resolvedAt": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CatalogItem": {
        "type": "object",
        "required": [
          "id",
          "name",
          "stationcount"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/EntityId"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "stationcount": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}